:recycle: refactor : move BaseEntity
